requirements.txt
Django==5.1.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2

